+++ src/pages/LoginPage.tsx
@@ -1,8 +1,9 @@
-import React, { useState } from 'react';s
+import React, { useState } from 'react';
 import styled from '@emotion/styled';
 import { useNavigate, useLocation } from 'react-router-dom';
 import { Navbar } from '@/components/Navbar';
 import { Layout } from '@/components/Layout';
+import { LoginForm } from './LoginForm';
 const LoginWrapper = styled.main`
   
 
@@ -62,41 +63,21 @@ export default function LoginPage() {
   const location = useLocation();
   const fromPath = location.state?.from?.pathname || '/';
 
-  const [id, setId] = useState('');
-  const [pw, setPw] = useState('');
-  const [idError, setIdError] = useState('');
-  const [pwError, setPwError] = useState('');
-
-  const handleIdChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setId(e.target.value);
-  };
-
-  const handlePwChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setPw(e.target.value);
-  };
-
-  const handleIdBlur = () => {
-    if (!id.trim()) {
-      setIdError('ID를 입력해주세요.');
-    } else if (!/^[^\s@]+@[^\s@]+\.[^\s@]+$/.test(id)) {
-      setIdError('ID는 이메일 형식으로 입력해주세요.');
-    } else {
-      setIdError('');
-    }
-  };
-
-  const handlePwBlur = () => {
-    if (!pw.trim()) {
-      setPwError('PW를 입력해주세요.');
-    } else if (pw.length < 8) {
-      setPwError('PW는 최소 8글자 이상이어야 합니다.');
-    } else {
-      setPwError('');
-    }
-  };
+  const {
+    id,
+    pw,
+    idError,
+    pwError,
+    handleIdChange,
+    handlePwChange,
+    handleIdBlur,
+    handlePwBlur,
+    isValid,
+  } = LoginForm();
 
   const handleSubmit = (e: React.FormEvent<HTMLFormElement>) => {
     e.preventDefault();
+    if (!isValid) return;
     navigate(fromPath, { replace: true });
   };
 
@@ -124,7 +105,9 @@ export default function LoginPage() {
               required
             />
             {pwError && <LoginErrorMsg>{pwError}</LoginErrorMsg>}
-            <LoginButton type="submit">로그인</LoginButton>
+            <LoginButton type="submit" disabled={!isValid}>
+              로그인
+            </LoginButton>
           </form>
         </LoginWrapper>
       </Layout>
